// frontend/src/api/models/SubtitleStyle.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SubtitleStyle = {
    DeepDiver: 'deep_diver',
    Karaoke: 'karaoke',
    Beasty: 'beasty',
    Default: 'default',
    RoundedBox: 'rounded_box',
    MessageBox: 'message_box',
    RectanglePerWordPopup: 'rectangle_per_word_popup',
    KaraokePopupRectangle: 'karaoke_popup_rectangle',
    Regular: 'regular',
    RectanglePerWord: 'rectangle_per_word',
    MessageBubble: 'message_bubble'
} as const;
export type SubtitleStyle = typeof SubtitleStyle[keyof typeof SubtitleStyle];


export function instanceOfSubtitleStyle(value: any): boolean {
    for (const key in SubtitleStyle) {
        if (Object.prototype.hasOwnProperty.call(SubtitleStyle, key)) {
            if (SubtitleStyle[key as keyof typeof SubtitleStyle] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SubtitleStyleFromJSON(json: any): SubtitleStyle {
    return SubtitleStyleFromJSONTyped(json, false);
}

export function SubtitleStyleFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitleStyle {
    return json as SubtitleStyle;
}

export function SubtitleStyleToJSON(value?: SubtitleStyle | null): any {
    return value as any;
}

export function SubtitleStyleToJSONTyped(value: any, ignoreDiscriminator: boolean): SubtitleStyle {
    return value as SubtitleStyle;
}